backend/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==3.2.15
django-extensions==3.2.1
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dotenv-python==0.0.1
drf-extra-fields==3.4.0
Faker==12.0.1
flake8==5.0.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mixer==7.1.2
mypy==0.971
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==23.0
Pillow==9.2.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
